Fix Makefile
The previous short and pleasant docker build piped to the tee had to be
rewritten into much more convoluted script 'tools/docker-build.sh' to
handle better the errors during the docker build.

In another words - the previous simple command in the Makefile:
    % docker build ... | tee logfile

had a flaw...when docker build failed the make command would happily
continue and the reason for that is the tee command which hid the
previous erroneous result code from the previous command in the pipe.

Now, to remedy the situation while providing the original behaviour
there is a new script:

    tools/docker-build.sh

Just by looking into it one can see that this could not be written in
the Makefile directly...

Signed-off-by: Petr Ospalý <petr@ospalax.cz>