Add initial support for ONE NFV appliance lease
I am still not completely happy with this, but it works and it assigns
ip address matching the hw address as requested by OpenNebula for NFV
appliance.

I can see some corner cases when it can behave shady (when lease was
already provided but it contradicts the hook requirement etc.).

Also in this particular case it was not needed BUT I could not find any
reliable way how to check that my "random" lease was not already taken.
Unfortunately Kea server has very simple address allocator which just
tracks the last assigned address and it just iterates to higher one till
it reaches the maximum for the particular pool...

Signed-off-by: Petr Ospalý <petr@ospalax.cz>