Improve Kea config generation
This commit explicitly states the interface for the subnet lease. If it
is unnecessary or it fixes some issues remains to be seen.

Kea does not provide any explanation or documentation on this behaviour
and usage of 'interface' field in 'subnet4' section can be found without
explanation only in one occurrence in documentation:
https://kea.readthedocs.io/en/v1_6_0/arm/dhcp4-srv.html#host-reservation-in-dhcpv4

Signed-off-by: Petr Ospalý <petr@ospalax.cz>